+++ apps/toropass_wallet/lib/features/home/data/models/oauth_permission_request_model.dart
@@ -5,6 +5,8 @@ class OAuthPermissionRequestModel extends OAuthPermissionRequestEntity {
     required super.clientId,
     required super.redirectUri,
     required super.appName,
+    required super.codeChallenge,
+    required super.codeChallengeMethod,
     super.state,
     super.scopes,
   });
@@ -35,6 +37,8 @@ class OAuthPermissionRequestModel extends OAuthPermissionRequestEntity {
           'Connected App',
       state: query['state']?.trim(),
       scopes: scopes,
+      codeChallenge: query['code_challenge']?.trim() ?? '',
+      codeChallengeMethod: query['code_challenge_method']?.trim() ?? '',
     );
   }
 }