+++ apps/toropass_wallet/lib/features/home/presentation/provider/permission_notifier.dart
@@ -41,6 +41,8 @@ class PermissionNotifier extends _$PermissionNotifier {
           clientId: request.clientId,
           redirectUri: request.redirectUri,
           scopes: request.scopes,
+          codeChallenge: request.codeChallenge,
+          codeChallengeMethod: request.codeChallengeMethod,
         ),
       );
       state = state.copyWith(authorizeState: response);